Adds transaction hooks feature.
With this PR, binding a datasource provides a new class TransactionHookManager that allows clients to add a callback that will be run only upon the successful commitment of the current transaction.

The algorithm I ended up implementing is not the same as the one we discussed earlier. Instead of using Hibernate-specific lifecycle callbacks, this one is a pure-JPA implementation that takes advantage of the fact that Onami Persist does manual transaction management. This was necessary because Onami Persist is itself pure-JPA and I didn't want to add a Hibernate-specific dependency to it, and anyway once I'd figured out how the code worked this way seemed about equivalently difficult.

Note that Onami Persist's implementation is pretty weird, and in particular everything ends up getting backed by a few singletons that use thread-local context to manage their state. I think this has to do with the way the JPA definition treats the notion of a transaction as a dynamically-scoped ambient property. In any case though please try to think hard about whether this code is thread-safe and deals with different datastores properly.